// ui/lib/api/schemas/projectStyle.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 */
import type { FontPolicy } from './fontPolicy';
import type { FontReviewQueueItem } from './fontReviewQueueItem';
import type { FontStyleProfile } from './fontStyleProfile';

export interface ProjectStyle {
  /** @nullable */
  defaultFont?: string | null;
  /** @nullable */
  fontProfile?: null | FontStyleProfile;
  fontPolicy?: FontPolicy;
  fontReviewQueue?: FontReviewQueueItem[];
}